moving the pickleDataset into ingest/labeled.py
